[Feature] Killing the browser
It's too easy to deactivate the add-on in private mode. So we're now detecting
if the addon is deactivated in private mode and killing the browser if that's
the case. Normally, each session should start with the addon enabled in private
mode, so this should not lock the user out of the browser entirely.